feat(medusa): add build versionizing, add to utils app

+++ medusa-nx-native-federation-example/packages/native-federation-plugin/src/executors/build/directory-util.ts
@@ -1,53 +1,67 @@
 import { basename, join } from 'path';
-import { readdir, lstatSync, rmSync, existsSync, mkdirSync, copyFileSync } from 'fs';
+import { lstatSync, rmSync, existsSync, mkdirSync, copyFileSync, readdirSync } from 'fs';
 
 /**
  *
  */
-export function cleanDirectoryFiles(directoryPath: string) {
-  readdir(directoryPath, (error: NodeJS.ErrnoException, files: string[]) => {
-    if (error) {
-      return;
-    }
-
-    for (const file of files) {
-      const filePath = join(directoryPath, file);
-
-      if (lstatSync(filePath).isFile()) {
-        rmSync(filePath, { recursive: true, force: true });
-      }
-    }
-  });
+export function isDirectory(directoryPath: string): boolean {
+  return lstatSync(directoryPath).isDirectory();
 }
 
 /**
  *
  */
-export function copyDirectory(fromPath: string, toPath: string) {
+export function cleanDirectory(directoryPath: string, ignoreFn?: (filePath: string) => boolean) {
+  for (const file of readdirSync(directoryPath)) {
+    const filePath = join(directoryPath, file);
+
+    if (ignoreFn && ignoreFn(filePath)) {
+      continue;
+    }
+
+    rmSync(filePath, { recursive: true, force: true });
+  }
+}
+
+/**
+ *
+ */
+export function copyDirectory(
+  fromPath: string,
+  toPath: string,
+  ignoreFn?: (sourcePath: string, outputPath: string) => boolean,
+  keepFolder = false
+) {
+  if (!existsSync(fromPath)) {
+    return;
+  }
+
   if (lstatSync(fromPath).isFile()) {
     copyFileSync(fromPath, join(toPath, basename(fromPath)));
     return;
+  }
+
+  if (keepFolder) {
+    toPath = join(toPath, basename(fromPath));
   }
 
   if (!existsSync(toPath)) {
     mkdirSync(toPath);
   }
 
-  readdir(fromPath, (error: NodeJS.ErrnoException, files: string[]) => {
-    if (error) {
-      return;
+  for (const file of readdirSync(fromPath)) {
+    const sourcePath = join(fromPath, file);
+    const outputPath = join(toPath, file);
+
+    if (ignoreFn && ignoreFn(sourcePath, outputPath)) {
+      continue;
     }
 
-    for (const file of files) {
-      const sourcePath = join(fromPath, file);
-      const outputPath = join(toPath, file);
-
-      if (lstatSync(sourcePath).isFile()) {
-        copyFileSync(sourcePath, outputPath);
-        continue;
-      }
-
-      copyDirectory(sourcePath, outputPath);
+    if (lstatSync(sourcePath).isFile()) {
+      copyFileSync(sourcePath, outputPath);
+      continue;
     }
-  });
+
+    copyDirectory(sourcePath, outputPath, ignoreFn, keepFolder);
+  }
 }
